// pocketsync-web/api-client/model/device-change-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PocketSync API
 * API documentation for PocketSync backend services
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DeviceChangeDto } from './device-change-dto';

/**
 * 
 * @export
 * @interface DeviceChangeResponseDto
 */
export interface DeviceChangeResponseDto {
    /**
     * 
     * @type {Array<DeviceChangeDto>}
     * @memberof DeviceChangeResponseDto
     */
    'items': Array<DeviceChangeDto>;
    /**
     * 
     * @type {number}
     * @memberof DeviceChangeResponseDto
     */
    'total': number;
    /**
     * 
     * @type {number}
     * @memberof DeviceChangeResponseDto
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof DeviceChangeResponseDto
     */
    'limit': number;
    /**
     * 
     * @type {number}
     * @memberof DeviceChangeResponseDto
     */
    'pages': number;
}