// pocketsync-web/api-client/model/device-change-timeline-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PocketSync API
 * API documentation for PocketSync backend services
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DeviceChangeDto } from './device-change-dto';

/**
 * 
 * @export
 * @interface DeviceChangeTimelineDto
 */
export interface DeviceChangeTimelineDto {
    /**
     * Record identifier
     * @type {string}
     * @memberof DeviceChangeTimelineDto
     */
    'recordId': string;
    /**
     * Table name
     * @type {string}
     * @memberof DeviceChangeTimelineDto
     */
    'tableName': string;
    /**
     * 
     * @type {Array<DeviceChangeDto>}
     * @memberof DeviceChangeTimelineDto
     */
    'changes': Array<DeviceChangeDto>;
}